requirements.txt
flask
flask-cors
python-dotenv
crewai
requests
